# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login/google:
    get:
      operationId: authLoginGoogle
      tags: [auth]
      summary: Get Google login URL
      parameters:
        - name: redirect
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Google OAuth URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url
...
  /auth/google:
    get:
      operationId: getGoogleAuthUrl
      tags: [sheets]
      summary: Get Google Sheets OAuth URL
      parameters:
        - name: redirect
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Google OAuth URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url
...